+++ app/javascript/vue/tasks/matrix_row_coder/store/actions/createObservation.js
@@ -30,7 +30,7 @@ export default function ({ commit, state }, args) {
       })
 
       commit(MutationNames.SetDescriptorSavedOnce, args.descriptorId)
-      if (isValidResponseData(responseData)) { commit(MutationNames.SetObservationId, makeObservationIdArgs(responseData.id, responseData.global_id)) }    
+      if (isValidResponseData(responseData)) { commit(MutationNames.SetObservationId, makeObservationIdArgs(responseData.id, responseData.global_id)) }
       return true
     }, response => {
       commit(MutationNames.SetDescriptorSaving, {
@@ -57,7 +57,7 @@ export default function ({ commit, state }, args) {
   }
 
   function setupContinuosPayload (payload) {
-    return Object.assign(payload, { 
+    return Object.assign(payload, {
       continuous_value: observation.continuousValue,
       continuous_unit: observation.continuousUnit
     })