+++ app/javascript/vue/tasks/matrix_row_coder/store/helpers/ComponentNames.js
@@ -1,8 +1,9 @@
 const ComponentNames = {
-  Qualitative: `qualitative-descriptor`,
-  Presence: `presence-descriptor`,
-  Continuous: `continuous-descriptor`,
-  Sample: `sample-descriptor`
+  Media: 'media-descriptor',
+  Qualitative: 'qualitative-descriptor',
+  Presence: 'presence-descriptor',
+  Continuous: 'continuous-descriptor',
+  Sample: 'sample-descriptor'
 }
 
 export default ComponentNames